chore: get rid of udr
