+++ internal/pcf/pcf.go
@@ -13,7 +13,6 @@ import (
 
 func Start(dbInstance *db.Database) error {
 	pcfCtx = &PCFContext{
-		TimeFormat:             "2006-01-02 15:04:05",
 		PcfSuppFeats:           make(map[models.ServiceName]openapi.SupportedFeature),
 		SessionRuleIDGenerator: idgenerator.NewGenerator(1, math.MaxInt64),
 		QoSDataIDGenerator:     idgenerator.NewGenerator(1, math.MaxInt64),
