+++ internal/udm/context.go
@@ -9,6 +9,7 @@ package udm
 import (
 	"sync"
 
+	"github.com/ellanetworks/core/internal/db"
 	"github.com/ellanetworks/core/internal/util/idgenerator"
 	"github.com/ellanetworks/core/internal/util/suci"
 	"github.com/omec-project/openapi"
@@ -26,6 +27,7 @@ const (
 )
 
 type UDMContext struct {
+	DbInstance                     *db.Database
 	NfId                           string
 	GroupId                        string
 	UriScheme                      models.UriScheme
@@ -36,6 +38,8 @@ type UDMContext struct {
 	SubscriptionOfSharedDataChange sync.Map                     // subscriptionID as key
 	SuciProfiles                   []suci.SuciProfile
 	EeSubscriptionIDGenerator      *idgenerator.IDGenerator
+	SdmSubscriptionIDGenerator     int
+	UESubsCollection               sync.Map // map[ueId]*UESubsData
 }
 
 type UdmNFContext struct {
