+++ internal/udm/udm.go
@@ -5,6 +5,7 @@ package udm
 import (
 	"math"
 
+	"github.com/ellanetworks/core/internal/db"
 	"github.com/ellanetworks/core/internal/util/idgenerator"
 	"github.com/ellanetworks/core/internal/util/suci"
 	"github.com/omec-project/openapi/models"
@@ -14,7 +15,7 @@ const (
 	UDM_HNP_PRIVATE_KEY = "c09c17bddf23357f614f492075b970d825767718114f59554ce2f345cf8c4b6a"
 )
 
-func Start() error {
+func Start(dbInstance *db.Database) error {
 	udmContext.UriScheme = models.UriScheme_HTTP
 	udmContext.SuciProfiles = []suci.SuciProfile{
 		{
@@ -24,5 +25,7 @@ func Start() error {
 	}
 	udmContext.NfService = make(map[models.ServiceName]models.NfService)
 	udmContext.EeSubscriptionIDGenerator = idgenerator.NewGenerator(1, math.MaxInt32)
+	udmContext.SdmSubscriptionIDGenerator = 1
+	udmContext.DbInstance = dbInstance
 	return nil
 }
