+++ internal/udm/ue_context.go
@@ -15,31 +15,21 @@ import (
 type UdmUeContext struct {
 	Supi                              string
 	Gpsi                              string
-	ExternalGroupID                   string
 	Nssai                             *models.Nssai
 	Amf3GppAccessRegistration         *models.Amf3GppAccessRegistration
-	AmfNon3GppAccessRegistration      *models.AmfNon3GppAccessRegistration
 	AccessAndMobilitySubscriptionData *models.AccessAndMobilitySubscriptionData
 	SmfSelSubsData                    *models.SmfSelectionSubscriptionData
 	UeCtxtInSmfData                   *models.UeContextInSmfData
-	TraceData                         *models.TraceData
 	SessionManagementSubsData         map[string]models.SessionManagementSubscriptionData
-	SubsDataSets                      *models.SubscriptionDataSets
 	SubscribeToNotifChange            map[string]*models.SdmSubscription
 	SubscribeToNotifSharedDataChange  *models.SdmSubscription
 	PduSessionID                      string
-	UdrUri                            string
-	UdmSubsToNotify                   map[string]*models.SubscriptionDataSubscriptions
-	EeSubscriptions                   map[string]*models.EeSubscription // subscriptionID as key
-	TraceDataResponse                 models.TraceDataResponse
 	amSubsDataLock                    sync.Mutex
 	smfSelSubsDataLock                sync.Mutex
 	SmSubsDataLock                    sync.RWMutex
 }
 
 func (ue *UdmUeContext) init() {
-	ue.UdmSubsToNotify = make(map[string]*models.SubscriptionDataSubscriptions)
-	ue.EeSubscriptions = make(map[string]*models.EeSubscription)
 	ue.SubscribeToNotifChange = make(map[string]*models.SdmSubscription)
 }
